backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
ollama==0.4.0
requests==2.32.0
pillow>=10.0.0
feedparser==6.0.11
python-multipart==0.0.20
pydantic>=2.0.0
python-dotenv==1.0.1
aiofiles==24.1.0
beautifulsoup4==4.12.3
